+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package pylon_ros2_camera
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.2 (2024-12-12)
+-------------------
+* Fixing bug related to ExposureTime/ExposureTimeAbs variable ID problem
+* Fixing compatibility problem with camera temperature value
+
 3.0.1 (2024-11-27)
 -------------------
 * Removing image_common as a submodule.